requirements.txt
amqp==5.1.1
asgiref==3.6.0
async-timeout==4.0.2
billiard==3.6.4.0
celery==5.2.7
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cron-descriptor==1.2.35
cryptography==40.0.1
defusedxml==0.7.1
diff-match-patch==20200713
Django==4.2
django-celery-beat==2.5.0
django-celery-results==2.5.0
django-ckeditor==6.5.1
django-dark==0.1.9
django-import-export==3.1.0
django-js-asset==2.0.0
django-modeltranslation==0.18.9
django-phonenumber-field==7.0.2
django-rosetta==0.9.8
django-timezone-field==5.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
et-xmlfile==1.1.0
idna==3.4
kombu==5.2.4
MarkupPy==1.14
oauthlib==3.2.2
odfpy==1.4.1
openpyxl==3.1.2
phonenumbers==8.13.8
Pillow==9.5.0
polib==1.2.0
postgres==4.0
prompt-toolkit==3.0.38
psycopg2-binary==2.9.5
psycopg2-pool==1.1
pycparser==2.21
PyJWT==2.6.0
python-crontab==2.7.1
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0
redis==4.5.4
requests==2.28.2
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==5.2.0
social-auth-core==4.4.1
sqlparse==0.4.3
tablib==3.4.0
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
vine==5.0.0
wcwidth==0.2.6
xlrd==2.0.1
xlwt==1.3.0
